% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\alias{bootstrap_ci.default}
\title{Generate bootstrapped CI}
\usage{
bootstrap_ci(
  data,
  max_coi = 25,
  seq_error = 0.01,
  coi_method = c("variant", "frequency"),
  solution_method = c("discrete", "continuous"),
  use_bins = FALSE,
  bin_size = 20,
  replicates = 100,
  parallel = TRUE,
  ncpus = 8
)

\method{bootstrap_ci}{default}(
  data,
  max_coi = 25,
  seq_error = 0.01,
  coi_method = c("variant", "frequency"),
  solution_method = c("discrete", "continuous"),
  use_bins = FALSE,
  bin_size = 20,
  replicates = 100,
  parallel = TRUE,
  ncpus = 8
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}

\item{solution_method}{Whether to estimate discrete or continuous COIs.}

\item{use_bins}{Do we calculate COI by comparing against the data grouped
into bins of changing \code{plaf} or not. Default = FALSE.}

\item{bin_size}{The minimum size of each bin of data.}

\item{replicates}{The number of bootstrap replicates.}

\item{parallel}{Whether to parallelize the confidence interval calculation.}

\item{ncpus}{The number of processes to be used in parallel operation.}
}
\description{
Generate bootstrapped confidence interval for COI estimates.
}
