% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi_testing.R
\name{coi_test}
\alias{coi_test}
\title{Test COIs}
\usage{
coi_test(
  repetitions = 10,
  COI = 3,
  COI_range = 5,
  PLAF = runif(1000, 0, 0.5),
  coverage = 100,
  alpha = 1,
  overdispersion = 0,
  epsilon = 0,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  method = c("end", "ideal", "overall"),
  dist_method = c("abs_sum", "sum_abs", "squared", "KL"),
  weighted = FALSE
)
}
\arguments{
\item{repetitions}{The number of times each sample will be run.}

\item{COI}{Complexity of infection.}

\item{COI_range}{A number indicating the range of COIs to compare the
simulated data to.}

\item{PLAF}{Vector of population-level allele frequencies at each locus.}

\item{coverage}{Coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters \code{p/overdispersion} and
\code{(1-p)/overdispersion}.}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. Applies in both directions.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{method}{The method to be employed. One of
\code{"end", "ideal", "overall"}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the "overall" method. One of
\code{"abs_sum", "sum_abs", "squared", "KL"}.}

\item{weighted}{An indicator indicating whether to compute weighted distance.}
}
\value{
A list of the following dataframes:
\describe{
  \item{\code{predicted_coi}:}{ A dataframe of the predicted COIs. COIs are
  predicted using \link{compute_coi}. Each column represents a separate set
  of parameters. Each row represents a predicted COI. Predictions are done
  many times, depending on the value of \code{repetitions}.}
  \item{\code{param_grid}:}{ The parameter grid. The parameter grid is all
  possible combinations of the parameters inputted. Each row represents a
  unique combination.}
  \item{\code{error_bias}:}{ A dataframe containing any parameter that was
  varied and the associated mean absolute error and bias (mean error). By
  showing only parameters that were varied, the output is easier to interpret
  and does not have information about parameters that were held constant.}
}
}
\description{
Runs several iterations of a full COI test with varying
parameters
}
