% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coiaf-package.R
\docType{package}
\name{coiaf-package}
\alias{coiaf}
\alias{coiaf-package}
\title{coiaf: Complexity of Infection Estimation with Allele Frequencies}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a direct method for estimating complexity of
    infection using within-sample allele frequencies from whole genome
    sequence data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bailey-lab.github.io/coiaf/}
  \item \url{https://github.com/bailey-lab/coiaf}
  \item Report bugs at \url{https://github.com/bailey-lab/coiaf/issues}
}

}
\author{
\strong{Maintainer}: Aris Paschalidis \email{aris_paschalidis@brown.edu} (\href{https://orcid.org/0000-0003-2247-1885}{ORCID})

Authors:
\itemize{
  \item OJ Watson \email{oj.watson@hotmail.co.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})
}

}
\keyword{internal}
