% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coiaf.R
\docType{package}
\name{coiaf}
\alias{coiaf}
\alias{coiaf-package}
\title{\pkg{coiaf} COI Estimation from Allele Frequency}
\description{
Provides a direct method for estimating complexity of infection
  using within-sample allele frequencies from whole genome sequence data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ojwatson.github.io/coiaf/}
  \item Report bugs at \url{https://github.com/ojwatson/coiaf/issues}
}

}
\author{
\strong{Maintainer}: OJ Watson \email{oj.watson@hotmail.co.uk} (\href{https://orcid.org/0000-0003-2374-0741}{ORCID})

Authors:
\itemize{
  \item Aris Paschalidis \email{aris_paschalidis@brown.edu} (\href{https://orcid.org/0000-0003-2247-1885}{ORCID})
}

}
