% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{compute_coi}
\alias{compute_coi}
\title{Compute COI}
\usage{
compute_coi(theory_cois, sim_coi, method = "end", cuts)
}
\arguments{
\item{theory_cois}{The theoretical COI curves to be examined}

\item{sim_coi}{The simulated coi curve}

\item{method}{The method to be employed.}

\item{cuts}{How often the data is summarized}
}
\value{
COI for the simulation
}
\description{
Generate the COI of the sample by comparing the simualated
coi curve to the theoretical coi curve.
}
