% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{compute_coi_regression}
\alias{compute_coi_regression}
\title{Compute COI based on residuals of all loci against theoretical curves}
\usage{
compute_coi_regression(
  data,
  data_type,
  max_coi = 25,
  seq_error = 0.01,
  distance = "squared",
  coi_method = "variant",
  seq_error_bin_size = 20
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim"} or \code{"real"}.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{distance}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; this function will solve a weighted least squares minimization
problem.}

\item{coi_method}{The method we will use to generate the theoretical
relationship. The method is either "variant" or "frequency". The default
value is "variant".}

\item{seq_error_bin_size}{Number of loci in smallest bin for estimating
sequence error}
}
\value{
A list of the following:
\itemize{
\item \code{coi}: The predicted COI of the sample.
\item \code{probability}: A probability density function representing the probability
of each COI.
}
}
\description{
Compute COI based on residuals of all loci against theoretical curves
}
