% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{distance_curves}
\alias{distance_curves}
\title{Compute distance between two curves}
\usage{
distance_curves(processed_data, theory_cois, distance = "squared")
}
\arguments{
\item{theory_cois}{The theoretical COI curves.}

\item{distance}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; this function will solve a weighted least squares minimization
problem.}
}
\value{
A list of the following:
\describe{
\item{\code{coi}}{The predicted COI for the simulation.}
\item{\code{dist}}{The distance between each theoretical COI and the
simulated COI curve.}
}
}
\description{
Compute the distance between two curves using several methods.
\describe{
\item{\code{abs_sum}}{Absolute value of sum of difference.}
\item{\code{sum_abs}}{Sum of absolute difference.}
\item{\code{squared}}{Sum of squared difference.}
}
}
\keyword{internal}
