% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood of a COI}
\usage{
likelihood(coi, processed_data, dist_method = "squared", weighted = TRUE)
}
\arguments{
\item{coi}{The COI for which the likelihood will be generated.}

\item{processed_data}{The processed COI data. This is the output of
\code{\link[=process_sim]{process_sim()}} or \code{\link[=process_real_data]{process_real_data()}}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
The likelihood for a specific COI value.
}
\description{
A function to generate the likelihood of a specific COI value.
The likelihood can be thought of the distance between two curves: the "real"
COI curve, generated from the inputted data, and the "simulated" COI curve,
which depends on the COI value specified.
}
