% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_coi}
\alias{optimize_coi}
\title{Optimize the COI}
\usage{
optimize_coi(
  data,
  data_type,
  max_coi = 25,
  seq_error = 0.01,
  bin_size = 20,
  distance = "squared",
  coi_method = "variant",
  use_bins = FALSE
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim"} or \code{"real"}.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{bin_size}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; to estimate the COI, all data points are used. Data points are
not grouped in bins of changing \code{plaf}.}

\item{distance}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; this function will solve a weighted least squares minimization
problem.}

\item{coi_method}{The method we will use to generate the theoretical
relationship. The method is either "variant" or "frequency". The default
value is "variant".}

\item{use_bins}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; to estimate the COI, all data points are used. Data points are
not grouped in bins of changing \code{plaf}.}
}
\value{
The predicted COI value.
}
\description{
A function to compute the COI of inputted data.
}
\details{
The function utilizes \code{\link[stats:optim]{stats::optim()}} In particular, the function utilizes
a quasi-Newton method to compute gradients and build a picture of the
surface to be optimized. The function uses a likelihood function as defined
by \code{\link[=likelihood]{likelihood()}}.
}
\seealso{
\code{\link[stats:optim]{stats::optim()}} for the complete documentation on the optimization
function.

Other optimization functions: 
\code{\link{likelihood}()},
\code{\link{optimize_coi_regression}()}
}
\concept{optimization functions}
