% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{optimize_coi_regression}
\alias{optimize_coi_regression}
\title{Compute COI based on all points fitted to best fitting curve for COI}
\usage{
optimize_coi_regression(
  data,
  data_type,
  max_coi = 25,
  seq_error = 0.01,
  distance = "squared",
  coi_method = "variant",
  seq_error_bin_size = 20
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim"} or \code{"real"}.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{distance}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}

\item{seq_error_bin_size}{Number of loci in smallest bin for estimating
sequence error}
}
\value{
The predicted COI value.
}
\description{
Compute COI based on all points fitted to best fitting curve for COI
}
\seealso{
\code{\link[stats:optim]{stats::optim()}} for the complete documentation on the optimization
function.

Other optimization functions: 
\code{\link{likelihood}()},
\code{\link{optimize_coi}()}
}
\concept{optimization functions}
