% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer}
\alias{optimizer}
\title{Compute COI Using Optimization}
\usage{
optimizer(
  data,
  data_type,
  max_coi = 25,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim} or \code{"real}.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum", "sum_abs", "squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
The COI value.
}
\description{
A function to compute the COI of inputted data.
The function utilizes the \code{\link[stats]{optim}} function. In particular,
the function utilizes a quasi-Newton method to compute gradients and build a
picture of the surface to be optimized.
}
\seealso{
\code{\link[stats]{optim}}
}
