% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot-simulation}
\alias{plot-simulation}
\alias{autoplot.sim}
\alias{plot.sim}
\title{Plot simulated data}
\usage{
\method{autoplot}{sim}(object, ...)

\method{plot}{sim}(x, ...)
}
\arguments{
\item{object, x}{An object of class \code{sim}. Derived from the output of
\code{\link[=sim_biallelic]{sim_biallelic()}}.}

\item{...}{Other arguments passed on to methods.}
}
\description{
Generate a simple plot visualizing simulated data. Compares the derived WSMAF
to the PLMAF.
}
\examples{
plot(sim_biallelic(coi = 2))
plot(sim_biallelic(coi = 5))
}
\seealso{
Other simulated data functions: 
\code{\link{process_sim}()},
\code{\link{sim_biallelic}()}
}
\concept{simulated data functions}
