% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_real}
\alias{process_real}
\title{Process real data}
\usage{
process_real(
  wsmaf,
  plmaf,
  coverage,
  seq_error = 0.01,
  bin_size = 20,
  coi_method = "variant"
)
}
\arguments{
\item{wsmaf}{The within-sample minor allele frequency.}

\item{plmaf}{The population-level minor allele frequency.}

\item{coverage}{The read coverage at each locus.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{bin_size}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; to estimate the COI, all data points are used. Data points are
not grouped in bins of changing \code{plaf}.}

\item{coi_method}{The method we will use to generate the theoretical
relationship. The method is either "variant" or "frequency". The default
value is "variant".}
}
\value{
A list of the following:
\itemize{
\item \code{data}: A tibble with
}
\itemize{
\item \code{plmaf_cut}: Breaks of the form \verb{[a, b)}.
\item \code{m_variant}: The average WSMAF or proportion of variant sites in each
segment defined by \code{plmaf_cut}.
\item \code{bucket_size}: The number of loci in each bucket.
\item \code{midpoints}: The midpoint of each bucket.
}
\itemize{
\item \code{seq_error}: The sequence error inferred.
\item \code{bin_size}: The minimum size of each bin.
\item \code{cuts}: The breaks utilized in splitting the data.
of each COI.
}
}
\description{
Generate the COI curve for real data.
}
\details{
The function computes whether a SNP is a variant site or not, based on the
WSMAF at that SNP. This process additionally accounts for potential
sequencing error.
}
\seealso{
\code{\link[=process_sim]{process_sim()}} to process simulated data.
}
\concept{real data functions}
