% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{process_sim}
\alias{process_sim}
\title{Process simulated data}
\usage{
process_sim(sim, seq_error = 0.01, cut = seq(0, 0.5, 0.01))
}
\arguments{
\item{sim}{Output of \code{\link[=sim_biallelic]{sim_biallelic()}}.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{How often the data is summarized.}
}
\value{
Simulated COI curve.
}
\description{
Generate the simulated COI curve.
}
\details{
Utilize the output of \code{\link[=sim_biallelic]{sim_biallelic()}}, which creates simulated
data. The PLAF is kept, and the function computes whether a SNP is a
variant site or not, based on the simulated WSAF at that SNP. This process
additionally accounts for potential sequencing error. To check whether the
simulated WSAF correctly indicated a variant site or not, the phased
haplotype of the parasites is computed.
}
\seealso{
\code{\link[=process_real]{process_real()}} to process real data.

Other simulated data functions: 
\code{\link{sim_biallelic}()}
}
\concept{simulated data functions}
