% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{process_simulated_coi}
\alias{process_simulated_coi}
\title{Process simulated data}
\usage{
process_simulated_coi(sim, seq_error = 0.01, cut = seq(0, 0.5, 0.01))
}
\arguments{
\item{sim}{Output of \code{\link[=sim_biallelic]{sim_biallelic()}}.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{How often the data is summarized.}
}
\value{
Simulated COI curve.
}
\description{
Generate the simulated COI curve. In order to do this, we
utilize the output of \code{\link[=sim_biallelic]{sim_biallelic()}}, which created simulated
data. We keep the PLAF, and compute whether a SNP is a variant or not, based
on the simulated WSAF at that SNP -- accounting for potential sequencing
error. To check whether our simulated WSAF correctly indicated a variant
site or not, we determine whether a site should be variant or not using
the phased haplotype of the parasites.
}
