% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coiaf.R
\docType{package}
\name{rdhs}
\alias{coiaf}
\alias{coiaf-package}
\alias{rdhs-package}
\title{\pkg{coiaf} COI estimation from allel frequency}
\description{
Provides a direct method for estimating complexity of infection
  using within-sample allele frequencies from whole genome sequence data.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/ojwatson/coiaf/issues}
}

}
\author{
\strong{Maintainer}: OJ Watson \email{oj.watson@hotmail.co.uk} (0000-0003-2374-0741)

Authors:
\itemize{
  \item Aris Paschalidis
}

}
