% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
rdirichlet(shape = rep(1, 3))
}
\arguments{
\item{shape}{Vector of shape parameters.}
}
\description{
Draw from a Dirichlet distribution given a vector of shape parameters. The
number of observations is inferred from the length of the shape vector.
}
\examples{
coiaf:::rdirichlet(c(1, 1, 1))
coiaf:::rdirichlet(c(1, 3, 2))
}
\seealso{
\code{\link[=rgamma]{rgamma()}} for additional details.

Other distributions: 
\code{\link{rbetabinom}()}
}
\concept{distributions}
\keyword{internal}
