% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_data.R
\name{run_real_data}
\alias{run_real_data}
\title{Run Real Data COIs}
\usage{
run_real_data(
  data,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  max_COI = 25,
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{data}{The dataset.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{max_COI}{A number indicating the maximum COI to compare the
simulated data to.}

\item{method}{The method to be employed. One of
\code{"end", "ideal", "overall"}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the "overall" method. One of
\code{"abs_sum", "sum_abs", "squared", "KL"}.}

\item{weighted}{An indicator indicating whether to compute weighted distance.}
}
\value{
A list of the following dataframes:
\describe{
  \item{\code{predicted_coi}}{A dataframe of the predicted COIs. COIs are
  predicted using \link{compute_coi}. Each column represents a separate set
  of parameters. Each row represents a predicted COI. Predictions are done
  many times, depending on the value of \code{repetitions}.}
  \item{\code{param_grid}}{The parameter grid. The parameter grid is all
  possible combinations of the parameters inputted. Each row represents a
  unique combination.}
  \item{\code{error_bias}}{A dataframe containing any parameter that was
  varied and the associated mean absolute error and bias (mean error). By
  showing only parameters that were varied, the output is easier to interpret
  and does not have information about parameters that were held constant.}
}
}
\description{
Run the algorithm on real data
}
