% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Plot sensitivity analysis}
\usage{
sensitivity_plot(data, dims, sub_title = NULL, title = NULL, caption = NULL)
}
\arguments{
\item{data}{The data to be plotted.}

\item{dims}{A list representing the number of rows and columns our plots
will be split into.}

\item{sub_title}{A list of titles for each individual subplot.}

\item{title}{The title of the overall figure.}

\item{caption}{The caption of the overall figure.}
}
\description{
The function is used to plot any sensitivity analysis that is
run. It takes in the output of \code{\link{coi_test}} and creates a grid of
plots.
}
