% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot_element}
\alias{sensitivity_plot_element}
\title{Helper function for \code{\link{sensitivity_plot}}}
\usage{
sensitivity_plot_element(data, loop_num, change_param, change_param_val)
}
\arguments{
\item{data}{The data to be plotted}

\item{loop_num}{The loop number}

\item{change_param}{The parameter being changed}

\item{change_param_val}{The values the changed parameter takes}
}
\value{
Plot
}
\description{
Plots a singular result
}
\keyword{internal}
