% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magenta.R
\name{sim_biallelic_magenta}
\alias{sim_biallelic_magenta}
\title{Simulate biallelic data from a magenta output}
\usage{
sim_biallelic_magenta(
  out,
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  epsilon = 0
)
}
\arguments{
\item{out}{Output of a \code{magenta} model.}

\item{coverage}{Coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters
\mjseqn{\frac{p}{overdispersion}} and
\mjseqn{\frac{1-p}{overdispersion}}.}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. Applies in both directions.}
}
\value{
A list of:
\itemize{
\item \code{coi}: The COI used to simulate the data.
\item \code{strain_proportions}: The strain proportion of each strain.
\item \code{phased}: The phased haplotype.
\item \code{data}: A dataframe of:
\itemize{
\item \code{plaf}: The population-level allele frequency.
\item \code{coverage}: The coverage at each locus.
\item \code{counts}: The count at each locus.
\item \code{wsaf}: The within-sample allele frequency.
}
\item \code{inputs}: A dataframe of function input arguments:
\itemize{
\item \code{alpha}: Shape parameters of Dirichlet controlling strain proportions.
\item \code{overdispersion}: Overdispersion in count data.
\item \code{relatedness}: Within sample relatedness between strains.
\item \code{epsilon}: Probability of a single read being miscalled.
}
}
}
\description{
Simulate biallelic data using a \code{magenta} model output.
}
\examples{
\dontrun{

 devtools::install_github("OJWatson/magenta")
 library(magenta)

 out <- magenta::pipeline(
 EIR=1,N=1000,years=2,num_loci = 1000,full_save=TRUE,
 drug_list = drug_list_create(drugs = list(magenta:::drug_create_dhappq()))
 )

 sims <- magenta_sim_format(out)

 sims_with_data <- sim_biallelic_magenta(sims[[1]])

}
}
\seealso{
Other simulated data functions: 
\code{\link{process_sim}()},
\code{\link{sim_biallelic}()}
}
\concept{simulated data functions}
