% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{simulated_coi}
\alias{simulated_coi}
\title{Generate Simulated COI Curve}
\usage{
simulated_coi(sim, seq_error, cuts)
}
\arguments{
\item{sim}{Output of \code{\link{sim_biallelic}}}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cuts}{How often the data is summarized.}
}
\value{
Simulated COI curve.
}
\description{
Generate the simulated COI curve. In order to do this, we
 utilize the output of \code{\link{sim_biallelic}}, which created simulated
 data. We keep the PLAF, and compute whether a SNP is a variant or not, based
 on the simulated WSAF at that SNP -- accounting for potential sequencing
 error. To check whether our simulated WSAF correctly indicated a variant
 site or not, we determine whether a site should be variant or not using
 the phased haplotype of the parasites.
}
