% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{single_theoretical_coi}
\alias{single_theoretical_coi}
\title{Single Theoretical COI}
\usage{
single_theoretical_coi(
  coi,
  plaf = seq(0, 0.5, l = 101),
  coi_method = "variant"
)
}
\arguments{
\item{coi}{The COI for which the curve will be calculated.}

\item{plaf}{The PLAF over which the curve will be calculated.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}
}
\value{
The theoretical COI curve for the specified PLAF.
}
\description{
Generate the theoretical COI curve for a particular COI value.
}
\details{
\loadmathjax
The theoretical curve can be visualized as the WSAF of an individual at
various PLAFs. At each locus, the theoretical COI is defined as:
\mjsdeqn{1-p^{coi}-q^{coi}}
}
\keyword{internal}
