% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{theoretical_coi}
\alias{theoretical_coi}
\title{Generate Theoretical COI Curves}
\usage{
theoretical_coi(cois, interval, method = "1")
}
\arguments{
\item{cois}{The COIs for which the curve will be calculated.}

\item{interval}{The interval over which the curve will be calculated.}

\item{method}{The method we will use to calculate the theoretical COI.
The method is either "1" or "2". The default value is "1".}
}
\value{
The theoretical COI curves for the specified COIs and interval.
}
\description{
Generate theoretical COI curves. At each locus defined as:
\eqn{1-p^coi-q^coi}. This expression is derived in the corresponding
manuscript.
}
