% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{world_map}
\alias{world_map}
\title{World map plot}
\usage{
world_map(data, variable, label = NULL, alpha = 0.1, breaks = c(1, 2))
}
\arguments{
\item{data}{The data to be plotted.}

\item{variable}{The variable the data will plot.}

\item{label}{The label for the variable.}

\item{alpha}{The alpha value for the plotted data.}

\item{breaks}{The breaks for the color scale.}
}
\description{
Plot a world map showing the COI in each region where reads were sampled
from.
}
\details{
Creates a world map and overlays the COI in each region. The magnitude of
the COI is indicated by both the color and the size of the bubble.
}
\seealso{
This \href{https://www.r-graph-gallery.com/bubble-map.html}{website} for
more information on creating bubble graphs in R.

Other plotting functions: 
\code{\link{error_plot}()},
\code{\link{sensitivity_plot}()}
}
\concept{plotting functions}
