% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\alias{bootstrap_ci.default}
\alias{bootstrap_ci.sim}
\title{Generate bootstrapped CI}
\usage{
bootstrap_ci(
  data,
  max_coi = 25,
  seq_error = 0.01,
  coi_method = c("variant", "frequency"),
  solution_method = c("discrete", "continuous"),
  use_bins = FALSE,
  bin_size = 20,
  replicates = 100,
  parallel = FALSE,
  ncpus = 8
)

\method{bootstrap_ci}{default}(
  data,
  max_coi = 25,
  seq_error = 0.01,
  coi_method = c("variant", "frequency"),
  solution_method = c("discrete", "continuous"),
  use_bins = FALSE,
  bin_size = 20,
  replicates = 100,
  parallel = FALSE,
  ncpus = 8
)

\method{bootstrap_ci}{sim}(
  data,
  max_coi = 25,
  seq_error = 0.01,
  coi_method = c("variant", "frequency"),
  solution_method = c("discrete", "continuous"),
  use_bins = FALSE,
  bin_size = 20,
  replicates = 100,
  parallel = FALSE,
  ncpus = 8
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}

\item{solution_method}{Whether to estimate discrete or continuous COIs.}

\item{use_bins}{Do we calculate COI by comparing against the data grouped
into bins of changing \code{plaf} or not. Default = FALSE.}

\item{bin_size}{The minimum size of each bin of data.}

\item{replicates}{The number of bootstrap replicates.}

\item{parallel}{Whether to parallelize the confidence interval calculation.
Note that parallelization only works on non-Windows machines.}

\item{ncpus}{The number of processes to be used in parallel operation.}
}
\value{
A \code{\link[tibble:tibble-package]{tibble()}} with columns:
\describe{
\item{coi}{The mean COI.}
\item{bias}{Bias of the statistic.}
\item{std.error}{The standard error of the statistic.}
\item{conf.low}{The lower 95\% confidence interval.}
\item{conf.high}{The upper 95\% confidence interval.}
}
}
\description{
Generate bootstrapped confidence interval for COI estimates.
}
\examples{
sim_data <- sim_biallelic(coi = 5, plmaf = runif(100, 0, 0.5))
bootstrap_ci(sim_data, solution_method = "continuous")
}
\seealso{
\code{\link[boot:boot]{boot::boot()}}, \code{\link[boot:boot.ci]{boot::boot.ci()}}, \code{\link[broom:tidy.boot]{broom::tidy.boot()}}
}
