% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi_testing.R
\name{coi_test}
\alias{coi_test}
\title{Test COIs}
\usage{
coi_test(
  repetitions = 10,
  COI = 3,
  max_COI = 25,
  PLAF = runif(1000, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  epsilon = 0,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{repetitions}{The number of times each sample will be run.}

\item{COI}{Complexity of infection.}

\item{max_COI}{A number indicating the maximum COI to compare the
simulated data to.}

\item{PLAF}{Vector of population-level allele frequencies at each locus.}

\item{coverage}{Coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters
\mjseqn{\frac{p}{overdispersion}} and
\mjseqn{\frac{1-p}{overdispersion}}.}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. Applies in both directions.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{method}{The method to be employed. One of \code{"end"}, \code{"ideal"},
\code{"overall"}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
A list of the following dataframes:
\itemize{
\item \code{predicted_coi}: A dataframe of the predicted COIs. COIs are
predicted using \code{\link[=compute_coi]{compute_coi()}}. Each column represents a separate set
of parameters. Each row represents a predicted COI. Predictions are done
many times, depending on the value of \code{repetitions}.
\item \code{probability}:A list of matrices containing the probability
that our model predicted each COI value. Each row contains the probability
for a different run. The first row contains the average probabilities over
all the runs.
\item \code{param_grid}: The parameter grid. The parameter grid is all
possible combinations of the parameters inputted. Each row represents a
unique combination.
\item \code{boot_error}: A dataframe containing information about the error
of the algorithm. The first column indicates the COI that was fed into the
simulation. The other columns indicate the mean absolute error (mae),
the lower and upper bounds of the 95\\% confidence interval and the bias.
}
}
\description{
Runs several iterations of a full COI test with varying parameters.
}
