% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{compute_coi}
\alias{compute_coi}
\title{Compute COI}
\usage{
compute_coi(sim, seq_error, cuts, theoretical_cois)
}
\arguments{
\item{sim}{Output of \code{\link{sim_biallelic}}.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cuts}{How often the data is summarized.}

\item{theoretical_cois}{The theoretical COI curves to be examined.}
}
\value{
Dataframe with averaged variant for several PLAF cuts.
}
\description{
Compute the COI of the sample. In order to do this, we utilize
 the output of \code{\link{sim_biallelic}}, which created simulated data.
 We keep the PLAF, and compute whether a SNP is a variant or not, based
 on the simulated WSAF at that SNP -- acounting for potential sequencing
 error. To check whether our simulated WSAF correctly indicated a variant
 site or not, we determine whether a site should be varaint or not using
 the phased haplotype of the parasites.
}
