% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{distance_curves}
\alias{distance_curves}
\title{Compute distance between two curves}
\usage{
distance_curves(
  theory_cois,
  sim_coi,
  cuts,
  dist_method = c("abs_sum", "sum_abs", "squared", "KL"),
  weighted = FALSE
)
}
\arguments{
\item{sim_coi}{The simulated COI curve, which is the output of
\link{sim_biallelic}.}

\item{cuts}{A vector indicating how often the data is summarized.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the "overall" method. One of
\code{"abs_sum", "sum_abs", "squared", "KL"}.}

\item{weighted}{An indicator indicating whether to compute weighted distance.}
}
\value{
The distance between two curves.
}
\description{
Compute the distance between two curves using several methods.
\describe{
  \item{\code{abs_sum}}{Absolute value of sum of difference.}
  \item{\code{sum_abs}}{Sum of absolute difference.}
  \item{\code{squared}}{Sum of squared difference.}
  \item{\code{KL}}{Kullback-Leibler divergence.}
  }
}
\keyword{internal}
