% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coi.R
\name{distance_curves}
\alias{distance_curves}
\title{Compute distance between two curves}
\usage{
distance_curves(
  processed_data,
  theory_cois,
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{processed_data}{The processed COI data. This is the output of
\code{\link[=process_sim]{process_sim()}} or \code{\link[=process_real]{process_real()}}.}

\item{theory_cois}{The theoretical COI curves.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
A list of the following:
\describe{
\item{\code{coi}}{The predicted COI for the simulation.}
\item{\code{dist}}{The distance between each theoretical COI and the
simulated COI curve.}
}
}
\description{
Compute the distance between two curves using several methods.
\describe{
\item{\code{abs_sum}}{Absolute value of sum of difference.}
\item{\code{sum_abs}}{Sum of absolute difference.}
\item{\code{squared}}{Sum of squared difference.}
}
}
\keyword{internal}
