% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{error_plot}
\alias{error_plot}
\title{Plot the mean absolute errors for a test}
\usage{
error_plot(
  data,
  fill = "COI",
  fill_levels = NULL,
  title = NULL,
  legend_title = fill,
  legend.position = "bottom",
  second_fill = NULL
)
}
\arguments{
\item{data}{The data to be plotted.}

\item{fill}{The variable the data will be separated by.}

\item{fill_levels}{The levels for the fill variable.}

\item{title}{The title of the plot. Default to \code{NULL}.}

\item{legend_title}{The text for the legend. Default to \code{NULL}.}

\item{legend.position}{The position of the legend. One of \code{"none",
"left", "right", "bottom", "top"}.}

\item{second_fill}{Indicates if there will be a second fill variable and
what it will be.}
}
\description{
Plots the mean absolute error and the confidence interval.
}
