% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood of a COI}
\usage{
likelihood(coi, processed_data, distance = "squared", coi_method = "variant")
}
\arguments{
\item{coi}{The COI for which the likelihood will be generated.}

\item{processed_data}{The processed COI data. This is the output of
\code{\link[=process_sim]{process_sim()}} or \code{\link[=process_real]{process_real()}}.}

\item{distance}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}
}
\value{
The likelihood for a specific COI value.
}
\description{
A function to generate the likelihood of a specific COI value.
}
\details{
The likelihood can be thought of the distance between two curves: the "real"
COI curve, generated from the inputted data, and the "simulated" COI curve,
which depends on the COI value specified. There are three different methods
implemented to compute the distance between two curves:
\itemize{
\item \code{abs_sum}: Absolute value of sum of difference.
\item \code{sum_abs}: Sum of absolute difference.
\item \code{squared}: Sum of squared difference.
}
}
\seealso{
Other optimization functions: 
\code{\link{optimize_coi}()}
}
\concept{optimization functions}
