% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_coi}
\alias{optimize_coi}
\title{Optimize the COI}
\usage{
optimize_coi(
  data,
  data_type,
  max_coi = 25,
  seq_error = NULL,
  bin_size = 20,
  distance = "squared",
  coi_method = "1"
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim"} or \code{"real"}.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{bin_size}{The minimum size of each bin of data.}

\item{distance}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "1" or "2". The default value is "1".}
}
\value{
The predicted COI value.
}
\description{
A function to compute the COI of inputted data.
}
\details{
The function utilizes \code{\link[stats:optim]{stats::optim()}} In particular, the function utilizes
a quasi-Newton method to compute gradients and build a picture of the
surface to be optimized. The function uses a likelihood function as defined
by \code{\link[=likelihood]{likelihood()}}.
}
\seealso{
\code{\link[stats:optim]{stats::optim()}} for the complete documentation on the optimization
function.

Other optimization functions: 
\code{\link{likelihood}()}
}
\concept{optimization functions}
