% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize_coi}
\alias{optimize_coi}
\title{Optimize the COI}
\usage{
optimize_coi(
  data,
  data_type,
  max_COI = 25,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{data}{The data for which the COI will be computed.}

\item{data_type}{The type of the data to be analyzed. One of
\code{"sim"} or \code{"real"}.}

\item{max_COI}{A number indicating the maximum COI to compare the
simulated data to.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
The predicted COI value.
}
\description{
A function to compute the COI of inputted data.
}
\details{
The function utilizes \code{\link[stats:optim]{stats::optim()}} In particular, the function utilizes
a quasi-Newton method to compute gradients and build a picture of the
surface to be optimized. The function uses a likelihood function as defined
by \code{\link[=likelihood]{likelihood()}}.
}
\seealso{
\code{\link[stats:optim]{stats::optim()}} for the complete documentation on the optimization
function.

Other optimization functions: 
\code{\link{likelihood}()}
}
\concept{optimization functions}
