% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process data}
\usage{
process(
  wsaf,
  plaf,
  seq_error = NULL,
  cut = seq(0, 0.5, 0.01),
  coi_method = "1"
)
}
\arguments{
\item{wsaf}{The within-sample allele frequency.}

\item{plaf}{The population-level allele frequency.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{cut}{How often the data is summarized.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "1" or "2". The default value is "1".}
}
\value{
Real COI curve.
}
\description{
Helper function to process data
}
\details{
The function computes whether a SNP is a variant site or not, based on the
simulated WSAF at that SNP. This process additionally accounts for potential
sequencing error.
}
\keyword{internal}
