% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_data.R
\name{process_real}
\alias{process_real}
\title{Generate Real COI Curve}
\usage{
process_real(wsaf, plaf, seq_error = 0.01, cut = seq(0, 0.5, 0.01))
}
\arguments{
\item{wsaf}{WSAF.}

\item{plaf}{PLAF.}

\item{seq_error}{The sequencing error.}

\item{cut}{How often the data is summarized.}
}
\value{
Real COI curve.
}
\description{
Generate the COI curve for real data.
}
