% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_sim}
\alias{process_sim}
\title{Process simulated data}
\usage{
process_sim(sim, seq_error = 0.01, bin_size = 20, coi_method = "variant")
}
\arguments{
\item{sim}{Output of \code{\link[=sim_biallelic]{sim_biallelic()}}.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{bin_size}{The minimum size of each bin of data.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}
}
\value{
A list of the following:
\itemize{
\item \code{data}: A tibble with
}
\itemize{
\item \code{plmaf_cut}: Breaks of the form \verb{[a, b)}.
\item \code{m_variant}: The average WSMAF or proportion of variant sites in each
segment defined by \code{plmaf_cut}.
\item \code{bucket_size}: The number of loci in each bucket.
\item \code{midpoints}: The midpoint of each bucket.
}
\itemize{
\item \code{seq_error}: The sequence error inferred.
\item \code{bin_size}: The minimum size of each bin.
\item \code{cuts}: The breaks utilized in splitting the data.
of each COI.
}
}
\description{
Generate the simulated COI curve.
}
\details{
Utilize the output of \code{\link[=sim_biallelic]{sim_biallelic()}}, which creates simulated
data. The PLMAF is kept, and the function computes whether a SNP is a
variant site or not, based on the simulated WSMAF at that SNP. This process
additionally accounts for potential sequencing error. To check whether the
simulated WSMAF correctly indicated a variant site or not, the phased
haplotype of the parasites is computed.
}
\seealso{
\code{\link[=process_real]{process_real()}} to process real data.

Other simulated data functions: 
\code{\link{plot-simulation}},
\code{\link{sim_biallelic}()}
}
\concept{simulated data functions}
