% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{rbetabinom}
\alias{rbetabinom}
\title{Draw from Beta-binomial distribution}
\usage{
rbetabinom(n = 1, k = 10, alpha = 1, beta = 1)
}
\arguments{
\item{n}{Number of draws.}

\item{k}{Number of binomial trials.}

\item{alpha}{First shape parameter of beta distribution.}

\item{beta}{Second shape parameter of beta distribution.}
}
\description{
Draw from Beta-binomial distribution.
}
\keyword{internal}
