% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Draw from Dirichlet distribution}
\usage{
rdirichlet(shape = rep(1, 3))
}
\arguments{
\item{shape}{vector of shape parameters.}
}
\description{
Draw from Dirichlet distribution given a vector of shape
  parameters.
}
