% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coi_testing.R
\name{run_coi_test}
\alias{run_coi_test}
\title{Run a single COI test}
\usage{
run_coi_test(
  COI = 3,
  max_COI = 25,
  PLAF = runif(1000, 0, 0.5),
  coverage = 100,
  alpha = 1,
  overdispersion = 0,
  epsilon = 0,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{COI}{Complexity of infection.}

\item{max_COI}{A number indicating the maximum COI to compare the
simulated data to.}

\item{PLAF}{Vector of population-level allele frequencies at each locus.}

\item{coverage}{Coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters \code{p/overdispersion} and
\code{(1-p)/overdispersion}.}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. Applies in both directions.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{method}{The method to be employed. One of
\code{"end", "ideal", "overall"}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum", "sum_abs", "squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}
}
\value{
Predicted COI value.
}
\description{
Runs a single full COI test.
}
\keyword{internal}
