% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_data.R
\name{run_real_data}
\alias{run_real_data}
\title{Run Real Data COIs}
\usage{
run_real_data(
  data,
  max_coi = 25,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  method = "overall",
  dist_method = "squared",
  weighted = TRUE
)
}
\arguments{
\item{data}{The dataset.}

\item{max_coi}{The maximum COI that the model will look at. Looks at the
theoretical COIs from 2:max_coi}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{method}{The method to be employed. One of
\code{"end", "ideal", "overall"}.}

\item{dist_method}{The distance method used to determine the distance between
the theoretical and simulated curves for the "overall" method. One of
\code{"abs_sum", "sum_abs", "squared", "KL"}.}

\item{weighted}{An indicator indicating whether to compute weighted distance.}
}
\value{
A list of the predicted cois and the probability distribution for the
predictions.
}
\description{
Run the algorithm on real data
}
