% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Function to plot results from \code{\link{coi_test}}}
\usage{
sensitivity_plot(data, plot_dims, change_param = NULL, change_param_val = NULL)
}
\arguments{
\item{data}{The data to be plotted}

\item{plot_dims}{A list representing the number of rows and columns}

\item{change_param}{The parameter being changed}

\item{change_param_val}{The values the changed parameter takes}
}
\value{
Plot
}
\description{
Plots a result
}
