% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{Sensitivity plot}
\usage{
sensitivity_plot(
  data,
  dims,
  result_type,
  sub_title = NULL,
  title = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{The data to be plotted.}

\item{dims}{A list representing the number of rows and columns our plots
will be split into.}

\item{result_type}{An indicator that indicates if a count or boxplot should
be plotted.}

\item{sub_title}{A list of titles for each individual subplot.}

\item{title}{The title of the overall figure.}

\item{caption}{The caption of the overall figure.}
}
\description{
Creates a plot of the sensitivity analysis.
}
\details{
\loadmathjax
The function takes in the output of \code{\link[=sensitivity]{sensitivity()}} and creates a grid of
plots. Each plot is created using \code{\link[ggplot2:geom_count]{ggplot2::geom_count()}}. The number of
observations at each location is counted and then the count is mapped to
point area on the plot.

The x-axis is the true COI, and the y-axis is the estimated COI. The counts
are plotted in blue, and red line is drawn with the equation \mjseqn{y = x}.
This line indicates where the blue circles should be if the algorithm was
100\% correct.
}
\seealso{
\code{\link[ggplot2:geom_count]{ggplot2::geom_count()}} for more information on count plots and the
\href{https://ggplot2.tidyverse.org/index.html}{ggplot2 website}.

Other plotting functions: 
\code{\link{error_plot}()},
\code{\link{world_map}()}
}
\concept{plotting functions}
