% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot_element}
\alias{sensitivity_plot_element}
\title{Single sensitivity plot}
\usage{
sensitivity_plot_element(data, loop_num, result_type, sub_title)
}
\arguments{
\item{data}{The data to be plotted.}

\item{loop_num}{The loop number. Represents how many total panels will
be plotted.}

\item{result_type}{An indicator that indicates if a count or boxplot should
be plotted.}

\item{sub_title}{A list of titles for each individual subplot.}
}
\description{
Creates a single plot of the sensitivity analysis. Used as a helper function
to \code{\link[=sensitivity_plot]{sensitivity_plot()}}.
}
\keyword{internal}
