% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{sensitivity_plot_element}
\alias{sensitivity_plot_element}
\title{Plot a single panel of the sensitivity analysis}
\usage{
sensitivity_plot_element(data, loop_num, change_param, change_param_val)
}
\arguments{
\item{data}{The data to be plotted.}

\item{loop_num}{The loop number. Represents how many total panels will
be plotted.}

\item{change_param}{The title of the plot. The title specifies the parameter
that is being changed.}

\item{change_param_val}{The values the changed parameter ranges over.}
}
\description{
Plots a single panel of the sensitivity analysis. Used as a
helper function to \code{\link{sensitivity_plot}}.
}
\keyword{internal}
