% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_biallelic}
\alias{sim_biallelic}
\title{Simulate biallelic data}
\usage{
sim_biallelic(
  coi,
  plaf = runif(10, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  relatedness = 0,
  epsilon = 0
)
}
\arguments{
\item{coi}{Complexity of infection.}

\item{plaf}{Vector of population-level allele frequencies at each locus.}

\item{coverage}{Coverage at each locus. If a single value is supplied then
the same coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-Binomially distributed, with
the beta distribution having shape parameters
\mjseqn{\frac{p}{overdispersion}} and
\mjseqn{\frac{1-p}{overdispersion}}.}

\item{relatedness}{The probability that a strain in mixed infections is
related to another. The implementation is similar to relatedness as defined
in THE REAL McCOIL simulations. In the \href{https://doi.org/10.1371/journal.pcbi.1005348}{original paper} this is defined as:
"... simulated relatedness (r) among lineages within the same host by
sampling alleles either from an existing lineage within the same host (with
probability r) or from the population (with probability (1-r))."}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. This error is applied in both directions.}
}
\value{
An object of class \code{sim}. Contains a list of
\link[tibble:tibble-package]{tibbles}:
\itemize{
\item \code{parameters} contains each parameter and the value used to simulate data.
\item \code{strain_proportions} contains the proportion of each strain.
\item \code{phased_haplotypes} contains the phased haplotype for each strain at each
locus.
\item \code{data} contains the following columns:
\itemize{
\item \code{plaf}: The population-level allele frequency.
\item \code{coverage}: The coverage at each locus.
\item \code{counts}: The count at each locus.
\item \code{wsaf}: The within-sample allele frequency.
}
}
}
\description{
Simulate biallelic data from a simple statistical model. Inputs
include the complexity of infection (COI), population-level allele
frequencies (PLAF), and some parameters dictating skew and error
distributions. Outputs include the phased haplotypes and the unphased read
count and coverage data.
}
\details{
\loadmathjax
Simulated data are drawn from a simple statistical model:
\enumerate{
\item Strain proportions are drawn from a symmetric Dirichlet
distribution with shape parameter \code{alpha}.
\item Phased haplotypes are drawn at every locus, one for each
\code{coi}. The allele at each locus is drawn from a Bernoulli
distribution with probability given by the \code{plaf}.
\item The "true" within-sample allele frequency at every locus is
obtained by multiplying haplotypes by their strain proportions, and
summing over haplotypes. Errors are introduced through the equation
\mjsdeqn{wsaf_{error} = wsaf(1-e) + (1-wsaf)e}
where \mjseqn{wsaf} is the WSAF without error and \mjseqn{e} is
the error parameter \code{epsilon}.
\item Final read counts are drawn from a beta-binomial distribution with
expectation \mjseqn{w_{error}}. The raw number of draws is given by the
\code{coverage}, and the skew of the distribution is given by the
\code{overdispersion} parameter. If the \code{overdispersion} is equal to
zero, then the distribution is binomial, rather than beta-binomial.
}
}
\examples{
sim_biallelic(coi = 5)
}
\seealso{
Other simulated data functions: 
\code{\link{plot-simulation}},
\code{\link{process_sim}()}
}
\concept{simulated data functions}
