% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{single_sensitivity}
\alias{single_sensitivity}
\title{Single sensitivity analysis}
\usage{
single_sensitivity(
  coi = 3,
  max_coi = 25,
  plaf = runif(1000, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  relatedness = 0,
  epsilon = 0,
  seq_error = 0.01,
  cut = seq(0, 0.5, 0.01),
  comparison = "overall",
  distance = "squared",
  weighted = TRUE,
  coi_method = "1"
)
}
\arguments{
\item{coi}{Complexity of infection.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{plaf}{Vector of population-level allele frequencies at each locus.}

\item{coverage}{Coverage at each locus. If a single value then the same
coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-binomially distributed, with
the beta distribution having shape parameters
\mjseqn{\frac{p}{overdispersion}} and
\mjseqn{\frac{1-p}{overdispersion}}.}

\item{relatedness}{The probability that a strain in mixed infections is
related to another. Default = 0 (unrelated). The implementation is similar
to relatedness as defined in THE REAL McCOIL simulations. In the original
paper (https://doi.org/10.1371/journal.pcbi.1005348) this is defined as:
"... simulated relatedness (r) among lineages within the same host by
sampling alleles either from an existing lineage within the same host
(with probability r) or from the population (with probability (1-r))."}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. Applies in both directions.}

\item{seq_error}{The level of sequencing error that is assumed.}

\item{cut}{A vector indicating how often the data is summarized.}

\item{comparison}{The method to be employed. One of \code{"end"}, \code{"ideal"},
\code{"overall"}.}

\item{distance}{The distance method used to determine the distance between
the theoretical and simulated curves for the \code{"overall"} method. One of
\code{"abs_sum"}, \code{"sum_abs"}, \code{"squared"}.}

\item{weighted}{An indicator indicating whether to compute the weighted
distance.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "1" or "2". The default value is "1".}
}
\value{
Predicted COI value.
}
\description{
Runs a single full COI sensitivity analysis.
}
\keyword{internal}
