% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{single_sensitivity}
\alias{single_sensitivity}
\title{Single sensitivity analysis}
\usage{
single_sensitivity(
  disc_or_cont,
  coi = 3,
  max_coi = 25,
  plmaf = runif(1000, 0, 0.5),
  coverage = 200,
  alpha = 1,
  overdispersion = 0,
  relatedness = 0,
  epsilon = 0,
  seq_error = 0.01,
  bin_size = 20,
  comparison = "overall",
  distance = "squared",
  coi_method = "variant",
  use_bins = FALSE
)
}
\arguments{
\item{disc_or_cont}{Whether to run a discrete or continuous COI estimation.}

\item{coi}{Complexity of infection.}

\item{max_coi}{A number indicating the maximum COI to compare the
simulated data to.}

\item{plmaf}{Vector of population-level minor allele frequencies at each
locus.}

\item{coverage}{Coverage at each locus. If a single value is supplied then
the same coverage is applied over all loci.}

\item{alpha}{Shape parameter of the symmetric Dirichlet prior on strain
proportions.}

\item{overdispersion}{The extent to which counts are over-dispersed relative
to the binomial distribution. Counts are Beta-Binomially distributed, with
the beta distribution having shape parameters
\mjseqn{\frac{p}{overdispersion}} and
\mjseqn{\frac{1-p}{overdispersion}}.}

\item{relatedness}{The probability that a strain in mixed infections is
related to another. The implementation is similar to relatedness as defined
in THE REAL McCOIL simulations. In the \href{https://doi.org/10.1371/journal.pcbi.1005348}{original paper} this is defined as:
"... simulated relatedness (r) among lineages within the same host by
sampling alleles either from an existing lineage within the same host (with
probability r) or from the population (with probability (1-r))."}

\item{epsilon}{The probability of a single read being miscalled as the other
allele. This error is applied in both directions.}

\item{seq_error}{The level of sequencing error that is assumed. If no value
is inputted, then we infer the level of sequence error.}

\item{bin_size}{The minimum size of each bin of data.}

\item{comparison}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no
longer supported; this function will compare the theoretical curve and
sample curve for all PLMAFs.}

\item{distance}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; this function will solve a weighted least squares minimization
problem.}

\item{coi_method}{The method we will use to generate the theoretical
relationship. The method is either "variant" or "frequency". The default
value is "variant".}

\item{use_bins}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This argument is no longer
supported; to estimate the COI, all data points are used. Data points are
not grouped in bins of changing \code{plaf}.}
}
\value{
Predicted COI value.
}
\description{
Runs a single full COI sensitivity analysis.
}
\keyword{internal}
