% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{single_theoretical_coi}
\alias{single_theoretical_coi}
\title{Generate Theoretical COI Curve}
\usage{
single_theoretical_coi(coi, interval)
}
\arguments{
\item{coi}{The COI}

\item{interval}{The interval over which the curve will be calculated}
}
\value{
The theoretical COI curve for the specified interval
}
\description{
Generate theoretical COI curve. At each locus defined as:
\eqn{1-p^coi-q^coi}. This expression is derived in the corresponding
manuscript.
}
\keyword{internal}
