% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{theoretical_coi}
\alias{theoretical_coi}
\title{Generate Theoretical COI Curves}
\usage{
theoretical_coi(cois, interval)
}
\arguments{
\item{cois}{The COIs for which the curve will be calculated}

\item{interval}{The interval over which the curve will be calculated}
}
\value{
The theoretical COI curves for the specified COIs and interval
}
\description{
Generate theoretical COI curves. At each locus defined as:
\eqn{1-p^coi-q^coi}. This expression is derived in the corresponding
manuscript.
}
