% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{theoretical_coi}
\alias{theoretical_coi}
\title{Theoretical COI}
\usage{
theoretical_coi(
  coi_range,
  plmaf = seq(0, 0.5, length.out = 101),
  coi_method = c("variant", "frequency")
)
}
\arguments{
\item{coi_range}{The COIs for which the relationship will be generated.}

\item{plmaf}{The population-level minor allele frequency over which the
relationship will be generated.}

\item{coi_method}{The method we will use to generate the theoretical
relationship. The method is either "variant" or "frequency". The default
value is "variant".}
}
\value{
A \code{\link[tibble:tibble-package]{tibble()}} containing the generated values. Each
column is named with the COI used. The last column of the tibble contains the
PLMAF.
}
\description{
\loadmathjax
Generate the theoretical relationship between the WSMAF (\mjseqn{\bf{w}}),
the PLMAF (\mjseqn{\bf{p}}), and the COI (\mjseqn{k}).
}
\examples{
theoretical_coi(1:5)
theoretical_coi(1:5, coi_method = "frequency")
}
