% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretical_coi.R
\name{theoretical_coi}
\alias{theoretical_coi}
\title{Theoretical COI}
\usage{
theoretical_coi(
  coi_range,
  plaf = seq(0, 0.5, length.out = 101),
  coi_method = "variant"
)
}
\arguments{
\item{coi_range}{The COIs for which the curve will be calculated.}

\item{plaf}{The PLAF over which the curve will be calculated.}

\item{coi_method}{The method we will use to calculate the theoretical COI.
The method is either "variant" or "frequency". The default value is
"variant".}
}
\value{
The theoretical COI curves for the specified COIs and PLAF.
}
\description{
Generate theoretical COI curves.
}
\details{
\loadmathjax
The theoretical curve can be visualized as the WSAF of an individual at
various PLAFs. At each locus, the theoretical COI is defined as:
\mjsdeqn{1-p^{coi}-q^{coi}} where \mjseqn{p} is the PLAF and \mjseqn{q} is
1 - PLAF.
}
