% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{world_map}
\alias{world_map}
\title{Plot a world map showing the COI}
\usage{
world_map(data, variable, label = NULL, alpha = 0.1, breaks = c(1, 2))
}
\arguments{
\item{data}{The data to be plotted.}

\item{variable}{The variable the data will plot.}

\item{label}{The label for the variable.}

\item{alpha}{The alpha value for the plotted data.}

\item{breaks}{The breaks for the color scale.}
}
\description{
Plot a world map showing the COI in each region where reads
were sampled from.
}
